Flash-tool/requirements.txt
intelhex==2.2.1
progressbar2==3.51.3
python-can==3.3.2
python-statemachine==0.8.0
